refactor(graphql-playground-html): merge the middleware package and the html package 👌
