+++ packages/graphql-playground-html/package.json
@@ -1,6 +1,6 @@
 {
   "name": "graphql-playground-html",
-  "version": "1.2.0",
+  "version": "1.2.1-beta.6",
   "homepage":
     "https://github.com/graphcool/graphql-playground/tree/master/packages/graphql-playground-html",
   "description":
@@ -26,5 +26,8 @@
   "typings": "dist/index.d.ts",
   "typescript": {
     "definition": "dist/index.d.ts"
+  },
+  "dependencies": {
+    "find-up": "^2.1.0"
   }
 }
