+++ packages/graphql-playground-middleware-express/src/index.ts
@@ -3,7 +3,7 @@ import {
   MiddlewareOptions,
   RenderPageOptions,
   renderPlaygroundPage,
-} from 'graphql-playground-middleware'
+} from 'graphql-playground-html'
 
 /* tslint:disable */
 const { version } = require('../package.json')
