--- packages/graphql-playground-middleware/.gitignore
@@ -1,18 +1,0 @@
-# See http://help.github.com/ignore-files/ for more about ignoring files.
-
-# dependencies
-node_modules
-example/node_modules
-example/yarn.lock
-
-# testing
-/coverage
-
-# production
-/dist
-
-# misc
-.DS_Store
-.env
-npm-debug.log
-.idea
