--- packages/graphql-playground-middleware/package.json
@@ -1,33 +1,0 @@
-{
-  "name": "graphql-playground-middleware",
-  "version": "1.2.1-beta.6",
-  "homepage":
-    "https://github.com/graphcool/graphql-playground/tree/master/packages/graphql-playground-middleware",
-  "description":
-    "GraphQL IDE for better development workflows (GraphQL Subscriptions, interactive docs & collaboration).",
-  "contributors": [
-    "Tim Suchanek <tim@graph.cool>",
-    "Johannes Schickling <johannes@graph.cool>",
-    "Mohammad Rajabifard <mo.rajbi@gmail.com>"
-  ],
-  "repository": "http://github.com/graphcool/graphql-playground.git",
-  "license": "SEE LICENSE IN LICENSE",
-  "main": "dist/index.js",
-  "files": ["dist"],
-  "scripts": {
-    "build": "tsc",
-    "prepublishOnly": "npm run build"
-  },
-  "keywords": ["graphql", "graphiql", "playground", "graphcool"],
-  "devDependencies": {
-    "@types/node": "^8.0.47",
-    "typescript": "^2.6.1"
-  },
-  "typings": "dist/index.d.ts",
-  "typescript": {
-    "definition": "dist/index.d.ts"
-  },
-  "dependencies": {
-    "find-up": "^2.1.0"
-  }
-}
