--- packages/graphql-playground-middleware/src/index.ts
@@ -1,5 +1,0 @@
-export {
-  renderPlaygroundPage,
-  MiddlewareOptions,
-  RenderPageOptions,
-} from './render-playground-page'
